Gbfs: Only show recently updated bikes
This changes the gbfs feed of available bikes to include only bikes that
have reported an update within the last hour.

Bikes check whether to report in every 3-8 minutes, and will decide
against reporting up to 5 times in a row if they believe their position
to be unchanged. That means that bikes should check in at least once
within 6 * 8 (= 48) minutes.

TODO: This should be globally configurable instead of hard-coded.

TODO: It would be great if the trackers would monitor their battery
      level and try to fire off a hail mary update before dying.